+++ crates/test/src/rt/mod.rs
@@ -112,9 +112,15 @@ use wasm_bindgen_futures::future_to_promise;
 const CONCURRENCY: usize = 1;
 
 pub mod browser;
+pub mod criterion;
 pub mod detect;
 pub mod node;
 mod scoped_tls;
+/// Directly depending on wasm-bindgen-test-based libraries should be avoided,
+/// as it creates a circular dependency that breaks their usage within `wasm-bindgen-test`.
+///
+/// Let's copy web-time.
+pub(crate) mod web_time;
 pub mod worker;
 
 /// Runtime test harness support instantiated in JS.
@@ -127,6 +133,9 @@ pub struct Context {
 }
 
 struct State {
+    /// In Benchmark
+    is_bench: bool,
+
     /// Include ignored tests.
     include_ignored: Cell<bool>,
 
@@ -226,7 +235,11 @@ trait Formatter {
     fn writeln(&self, line: &str);
 
     /// Log the result of a test, either passing or failing.
-    fn log_test(&self, name: &str, result: &TestResult);
+    fn log_test(&self, is_bench: bool, name: &str, result: &TestResult) {
+        if !is_bench {
+            self.writeln(&format!("test {} ... {}", name, result));
+        }
+    }
 
     /// Convert a thrown value into a string, using platform-specific apis
     /// perhaps to turn the error into a string.
@@ -278,7 +291,7 @@ impl Context {
     /// coordinated, and this will collect output and results for all executed
     /// tests.
     #[wasm_bindgen(constructor)]
-    pub fn new() -> Context {
+    pub fn new(is_bench: bool) -> Context {
         fn panic_handling(mut message: String) {
             let should_panic = CURRENT_OUTPUT.with(|output| {
                 let mut output = output.borrow_mut();
@@ -317,15 +330,11 @@ impl Context {
                 panic_handling(panic_info.to_string());
             }));
         });
-        #[cfg(all(
-            not(feature = "std"),
-            target_arch = "wasm32",
-            any(target_os = "unknown", target_os = "none")
-        ))]
+        #[cfg(not(feature = "std"))]
         #[panic_handler]
         fn panic_handler(panic_info: &core::panic::PanicInfo<'_>) -> ! {
             panic_handling(panic_info.to_string());
-            core::arch::wasm32::unreachable();
+            unreachable!();
         }
 
         let formatter = match detect::detect() {
@@ -338,6 +347,7 @@ impl Context {
 
         Context {
             state: Rc::new(State {
+                is_bench,
                 include_ignored: Default::default(),
                 failures: Default::default(),
                 succeeded_count: Default::default(),
@@ -371,10 +381,12 @@ impl Context {
     /// The promise returned resolves to either `true` if all tests passed or
     /// `false` if at least one test failed.
     pub fn run(&self, tests: Vec<JsValue>) -> Promise {
-        let noun = if tests.len() == 1 { "test" } else { "tests" };
-        self.state
-            .formatter
-            .writeln(&format!("running {} {}", tests.len(), noun));
+        if !self.state.is_bench {
+            let noun = if tests.len() == 1 { "test" } else { "tests" };
+            self.state
+                .formatter
+                .writeln(&format!("running {} {}", tests.len(), noun));
+        }
 
         // Execute all our test functions through their Wasm shims (unclear how
         // to pass native function pointers around here). Each test will
@@ -525,14 +537,16 @@ impl Context {
         ignore: Option<Option<&'static str>>,
     ) {
         // Remove the crate name to mimic libtest more closely.
-        // This also removes our `__wbgt_` prefix and the `ignored` and `should_panic` modifiers.
+        // This also removes our `__wbgt_` or `__wbgb_` prefix and the `ignored` and `should_panic` modifiers.
         let name = name.split_once("::").unwrap().1;
 
         if let Some(ignore) = ignore {
             if !self.state.include_ignored.get() {
-                self.state
-                    .formatter
-                    .log_test(name, &TestResult::Ignored(ignore.map(str::to_owned)));
+                self.state.formatter.log_test(
+                    self.state.is_bench,
+                    name,
+                    &TestResult::Ignored(ignore.map(str::to_owned)),
+                );
                 let ignored = self.state.ignored_count.get();
                 self.state.ignored_count.set(ignored + 1);
                 return;
@@ -622,8 +636,11 @@ impl State {
             if let TestResult::Err(_e) = result {
                 if let Some(expected) = should_panic {
                     if !test.output.borrow().panic.contains(expected) {
-                        self.formatter
-                            .log_test(&test.name, &TestResult::Err(JsValue::NULL));
+                        self.formatter.log_test(
+                            self.is_bench,
+                            &test.name,
+                            &TestResult::Err(JsValue::NULL),
+                        );
                         self.failures
                             .borrow_mut()
                             .push((test, Failure::ShouldPanicExpected));
@@ -631,17 +648,18 @@ impl State {
                     }
                 }
 
-                self.formatter.log_test(&test.name, &TestResult::Ok);
+                self.formatter
+                    .log_test(self.is_bench, &test.name, &TestResult::Ok);
                 self.succeeded_count.set(self.succeeded_count.get() + 1);
             } else {
                 self.formatter
-                    .log_test(&test.name, &TestResult::Err(JsValue::NULL));
+                    .log_test(self.is_bench, &test.name, &TestResult::Err(JsValue::NULL));
                 self.failures
                     .borrow_mut()
                     .push((test, Failure::ShouldPanic));
             }
         } else {
-            self.formatter.log_test(&test.name, &result);
+            self.formatter.log_test(self.is_bench, &test.name, &result);
 
             match result {
                 TestResult::Ok => self.succeeded_count.set(self.succeeded_count.get() + 1),